crescendo: add eigensolver class, fix bugs, output K and M
intermediate step in getting to a complete eigensolution.
Added an eigensolver class that is based off of one of the
Anasazi examples, but it relies on computing K^(-1)*M.  It
turns out that this doesn't work well for meshes with rigid
body modes since K^(-1) doesn't exist.

So, I added some non-negligible diagonal stiffness terms
to "K" so that I could test out the eigensolver, and it works
perfectly.  To verify, I built EpetraExt so that I could
output MatrixMarket files.  I read these into Octave and use
the sparse eigensolver in Octave.

There is an issue in the EigenSolver, it only seems to work
in Parallel.  I essentially copied an example and modified
it to work with different stiffness and mass matrices, and
in the process I simplified a few IFDEF lines that may have
handled the Serial/Parallel communication issues.  This will
have to be re-visited, but since I need to essentially
re-write the EigenSolver class to be able to handle rigid
body modes I am not going to worry about it right now.