crescendo: updates to EigenSolver, communicate variables across procs
This commit has a fully functional eigensolver with eigenvector
output to parallel exodus files.  It's really only been tested
with a 2-element hex problem, and right now any larger problems
just don't seem to converge.  This might be because parameters
for the eigensolver are hardcoded.  So in the future more appropriate
parameters for the eigensolver will be necessary.