eigensolver: merge Solve() and SolveIfpack() functions
Solve() used a shift and was the most tested option, but
did not employ a preconditioner.

SolveIfpack() used Ifpack to precondition but did not
have a shift.

Both options were liberally cloned from examples in the
Anasazi documentation, so this commit represents a merging
of the two examples.  Leaving preconditioner off by default
for now.

Still a lot of cleaning house to do in the Solve() routine ...