fix results output so it doesnt seg fault
still a lot of clean-up to do with indexing and how DOF
are mapped/handled.  essentially, need to keep track of DOF index
that corresponds to Epetra map along with which global node ID and
which local node ID (and/or index) it refers to.  Keeping this all
straight is critical to not messing things up.